UserInterfase almost completed

+++ src/GeoGem.Domain/Entities/Ticket.cs
@@ -5,6 +5,6 @@ namespace GeoGem.Domain.Entities;
 public class Ticket : Auditable
 {
     public long LandMarkId {  get; set; }
-    public int FlightDuration {  get; set; }
+    public double FlightDuration {  get; set; }
     public decimal Price { get; set; }
 }
