+++ src/GeoGem.Service/DTOs/Tickets/TicketForCreationDto.cs
@@ -3,6 +3,6 @@
 public class TicketForCreationDto
 {
     public long LandMarkId { get; set; }
-    public int FlightDuration { get; set; }
+    public double FlightDuration { get; set; }
     public decimal Price { get; set; }
 }
