+++ src/GeoGem.Service/DTOs/Tickets/TicketForResultDto.cs
@@ -4,6 +4,6 @@ public class TicketForResultDto
 {
     public long Id { get; set; }
     public long LandMarkId { get; set; }
-    public int FlightDuration { get; set; }
+    public double FlightDuration { get; set; }
     public decimal Price { get; set; }
 }
