+++ src/GeoGem.Service/DTOs/Tickets/TicketForUpdateDto.cs
@@ -4,6 +4,6 @@ public class TicketForUpdateDto
 {
     public long Id { get; set; }
     public long LandMarkId { get; set; }
-    public int FlightDuration { get; set; }
+    public double FlightDuration { get; set; }
     public decimal Price { get; set; }
 }
